Separate client component and adjust styling
